add null checks to test and move clients to appropriate folder
